+++ CHANGELOG.md
@@ -1,11 +1,18 @@
-rec_radiko CHANGELOG
+radiko CHANGELOG
 ====================
 
-This file is used to list changes made in each version of the rec_radiko cookbook.
+This file is used to list changes made in each version of the radiko cookbook.
+
+0.2.0
+-----
+- Changed cookbook name from rec_radiko to radiko
+- Set timezone - Asia/Tokyo
+- Added Ubuntu support
+- Divide recipe into install and record
 
 0.1.0
 -----
-- [your_name] - Initial release of rec_radiko
+- Initial release of radiko
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.